+++ modules/ai/mcp/build.gradle.kts
